+++ lib/src/get_state_manager/src/simple/get_view.dart
@@ -101,7 +101,7 @@ abstract class ReactiveGetView<T> extends StatelessWidget {
     if (useSmartFind) {
       return GetInstance().smartFind<T>(tag: tag)!;
     } else {
-      return GetInstance().find<T>(tag: tag);
+      return GetInstance().find<T>(tag: tag)!;
     }
   }